pyproject.toml
[tool.poetry]
name = "am-i-drunk"
version = "0.1.0"
description = "Image recognition based drunk level identifier for perties."
authors = ["Enrique <enrique.flores@utec.edu.pe>"]
readme = "README.md"

[tool.poetry.plugins."poetry.application.plugin"]
export = "poetry_plugin_export.plugins:ExportApplicationPlugin"

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^10.2.0"
ice = "^0.0.2"

[tool.poetry.group.dev.dependencies]
poethepoet = "^0.24.4"
pytest = "^7.4.3"
ruff = "^0.1.8"

[tool.poe.tasks]
test = "pytest"
lint = "ruff"
_build = "poetry build"
build = ["test", "_build"]
dev.script = "experiments.fifa_tesseract:main"
export = "poetry export --without-hashes --format=requirements.txt --all-extras --output requirements.txt"

[tool.ruff.lint]
select = ["ALL"]
ignore = ["D203", "D212", "FA", "ANN101"]

[tool.pyright]
typeCheckingMode = "strict"
pythonVersion = "3.11"
verboseOutput = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
